# garmin-stats/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garmin Stats — Local Health Dashboard API",
    "version": "2.0.0",
    "description": "Local-only REST API (binds 0.0.0.0) for the Holistic Runner App. Aggregates Garmin, Strava, and Withings data in a local SQLite DB.\n\nContract-hardening epic HRA-36 in progress: RFC 7807 `application/problem+json` error bodies with correct status codes (HRA-37 ✓), plural resource nouns incl. `/api/body-measurements` (HRA-39 ✓), the `{ data, page }` list envelope with limit/offset paging (HRA-38 ✓), PATCH (application/merge-patch+json) for partial settings updates (HRA-40 ✓), and the version prefix — every path is now under `/api/v1/...` (HRA-41 ✓) — have landed. The HRA-36 contract-hardening epic is complete. Drift between this file and the running API is treated as a bug."
  },
  "servers": [
    { "url": "http://127.0.0.1:3001", "description": "Local API server" }
  ],
...
  "components": {
    "securitySchemes": {
      "runsFreeSession": { "type": "apiKey", "in": "cookie", "name": "__Host-runsfree_session", "description": "Validated Runs Free session; private resources derive their owner from this credential, never from client input." }
    },
    "parameters": {
      "from": { "name": "from", "in": "query", "required": false, "schema": { "type": "string", "format": "date" }, "description": "Inclusive start date (YYYY-MM-DD). Defaults to 30 days ago." },
...
    },
    "responses": {
...
      "Unprocessable": { "description": "Request parsed but failed a validation rule.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
      "ServiceUnavailable": { "description": "The AI provider integration is not configured.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/login": {
      "get": { "tags": ["auth"], "summary": "Start the Auth0 Authorization Code BFF flow", "responses": { "302": { "description": "Redirects to the configured authorization endpoint." }, "503": { "$ref": "#/components/responses/ServiceUnavailable" } } }
    },
...
    "/api/v1/auth/session": {
      "get": { "tags": ["auth"], "summary": "Return server-authoritative session profile and the in-memory CSRF token", "security": [{ "runsFreeSession": [] }], "responses": { "200": { "description": "Current user and CSRF bootstrap.", "content": { "application/json": { "schema": { "type": "object", "required": ["user", "entitlements", "csrfToken"], "properties": { "user": { "type": "object" }, "entitlements": { "type": "array", "items": { "type": "string" } }, "csrfToken": { "type": "string" } } } } } }, "401": { "description": "Missing, expired, revoked, or unusable session." } } }
    },
    "/api/v1/auth/logout": {
      "post": { "tags": ["auth"], "summary": "Revoke the current opaque server session and clear its cookie", "security": [{ "runsFreeSession": [] }], "parameters": [{ "name": "X-RunsFree-CSRF", "in": "header", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Session cleared; the client returns to its signed-out bootstrap state." }, "401": { "description": "Missing or invalid session/CSRF proof." } } }
    },
    "/api/v1/account/profile": {
      "put": { "tags": ["account"], "summary": "Replace the caller's supported account profile fields", "security": [{ "runsFreeSession": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "display_name": { "type": ["string", "null"] }, "locale": { "type": ["string", "null"] }, "unit_system": { "type": ["string", "null"], "enum": ["metric", "imperial", null] }, "timezone": { "type": ["string", "null"] } } } } } }, "responses": { "200": { "description": "Sanitized profile; never exposes provider or session identifiers." }, "422": { "$ref": "#/components/responses/Unprocessable" } } }
    },
    "/api/v1/account/sessions/revoke-others": {
      "post": { "tags": ["account"], "summary": "Revoke every session except the current one", "description": "Requires a session issued within the previous five minutes; no session identifiers are returned.", "security": [{ "runsFreeSession": [] }], "responses": { "200": { "description": "Other sessions revoked." }, "401": { "description": "Missing or non-recent authentication." } } }
    },
    "/api/v1/account/exports": {
      "post": { "tags": ["account"], "summary": "Capture an owner-scoped personal-data export", "description": "The snapshot has a versioned manifest, excludes authentication/integration secrets and other users' data, and produces a secret-bound link expiring after 24 hours.", "security": [{ "runsFreeSession": [] }], "responses": { "201": { "description": "Expiring download link." } } }
    },
...
    "/api/v1/account/deletion-request": {
      "post": { "tags": ["account"], "summary": "Queue idempotent account deletion", "description": "Requires exact confirmation and a session issued within the previous five minutes. It immediately sets deletion_pending, revokes all sessions and integration credentials, and queues durable removal; retained backups are not immediately physically erased.", "security": [{ "runsFreeSession": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["confirmation"], "properties": { "confirmation": { "type": "string", "enum": ["DELETE MY ACCOUNT"] } } } } } }, "responses": { "202": { "description": "Deletion request queued." }, "401": { "description": "Missing or non-recent authentication." }, "422": { "$ref": "#/components/responses/Unprocessable" } } }
    },
...
    "/api/v1/activities": {
      "get": {
        "tags": ["activities"], "summary": "List activities in a date range (any source)",
        "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ],
        "responses": { "200": { "description": "Bare array of activities (no pagination envelope — see HRA-38).", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Activity" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } }
      },
      "delete": {
        "tags": ["activities"], "summary": "Soft-delete activities in a date range (moves to trash)",
        "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ],
        "responses": { "200": { "description": "Count soft-deleted.", "content": { "application/json": { "schema": { "type": "object", "properties": { "deleted": { "type": "integer" }, "from": { "type": "string" }, "to": { "type": "string" } } } } } }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/activities/count": {
      "get": { "tags": ["activities"], "summary": "Count activities in range", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Count.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Count" } } } } } }
    },
...
    "/api/v1/activities/races": {
      "get": { "tags": ["activities"], "summary": "Race-type activities (full history), for the 'link a race' dropdown", "parameters": [ { "$ref": "#/components/parameters/limit" }, { "$ref": "#/components/parameters/offset" } ], "responses": { "200": { "description": "Race activities.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/RaceActivity" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } }
    },
...
    "/api/v1/activities/confirm": {
      "post": { "tags": ["activities"], "summary": "Bulk-confirm classifications", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["ids"], "properties": { "ids": { "type": "array", "items": { "type": "integer" } }, "method": { "type": "string", "enum": ["ai", "statistical"], "default": "ai" } } } } } }, "responses": { "200": { "description": "Confirmed count.", "content": { "application/json": { "schema": { "type": "object", "properties": { "confirmed": { "type": "integer" } } } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "422": { "$ref": "#/components/responses/Unprocessable" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
    "/api/v1/activities/restore": {
      "post": { "tags": ["activities"], "summary": "Restore activities from trash", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["ids"], "properties": { "ids": { "type": "array", "items": { "type": "integer" } } } } } } }, "responses": { "200": { "description": "Restored count.", "content": { "application/json": { "schema": { "type": "object", "properties": { "restored": { "type": "integer" } } } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "422": { "$ref": "#/components/responses/Unprocessable" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
    "/api/v1/activities/purge": {
      "post": { "tags": ["activities"], "summary": "Permanently purge activities from trash", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["ids"], "properties": { "ids": { "type": "array", "items": { "type": "integer" } } } } } } }, "responses": { "200": { "description": "Purged count.", "content": { "application/json": { "schema": { "type": "object", "properties": { "purged": { "type": "integer" } } } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "422": { "$ref": "#/components/responses/Unprocessable" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
    "/api/v1/summary": {
      "get": { "tags": ["trends"], "summary": "Totals grouped by sport", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Per-sport summaries.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/SportSummary" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } }
    },
    "/api/v1/weekly": {
      "get": { "tags": ["trends"], "summary": "Weekly aggregates", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Weekly points.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/WeeklyPoint" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } }
    },
    "/api/v1/monthly": {
      "get": { "tags": ["trends"], "summary": "Monthly aggregates", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Monthly points.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/MonthlyPoint" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } }
    },
...
    "/api/v1/body-measurements": {
      "get": { "tags": ["body"], "summary": "Body measurements in range", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Measurements.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/BodyMeasurement" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } },
      "delete": { "tags": ["body"], "summary": "Soft-delete body measurements in a date range", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Count soft-deleted.", "content": { "application/json": { "schema": { "type": "object", "properties": { "deleted": { "type": "integer" }, "from": { "type": "string" }, "to": { "type": "string" } } } } } }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
    "/api/v1/body-measurements/count": {
      "get": { "tags": ["body"], "summary": "Count body measurements in range", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Count.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Count" } } } } } }
    },
    "/api/v1/body-measurements/monthly": {
      "get": { "tags": ["body"], "summary": "Monthly body averages", "parameters": [ { "$ref": "#/components/parameters/from" }, { "$ref": "#/components/parameters/to" } ], "responses": { "200": { "description": "Monthly body points.", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/BodyMonthlyPoint" } }, "page": { "$ref": "#/components/schemas/Page" } }, "required": ["data", "page"] } } } } } }
    },
...
    "/api/v1/settings/background-image": {
      "security": [{ "runsFreeSession": [] }],
      "get": { "tags": ["settings"], "summary": "Stream the current custom background image", "responses": { "200": { "description": "The image bytes.", "content": { "image/jpeg": {}, "image/png": {}, "image/webp": {}, "image/gif": {} } }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
...
  }
}